requirements.txt
telethon>=1.25.0
psycopg2-binary>=2.9.6
dbt-postgres>=1.7.0
ultralytics>=8.3.165
fastapi>=0.99.1
uvicorn[standard]>=0.22.0
dagster>=1.3.1
dagster-webserver>=1.3.1
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.32.3